[PimcoreX] introduce type declarations in CoreShop

+++ src/CoreShop/Bundle/CoreBundle/Validator/Constraints/AddToCartMinimumQuantityValidator.php
@@ -91,7 +91,7 @@ class AddToCartMinimumQuantityValidator extends ConstraintValidator
         }
     }
 
-    private function getExistingCartItemQuantityFromCart(OrderInterface $cart, OrderItemInterface $cartItem): int
+    private function getExistingCartItemQuantityFromCart(OrderInterface $cart, OrderItemInterface $cartItem): float
     {
         $product = $cartItem->getProduct();
         $quantity = 0;
